Fix :3 // 5.0

+++ fridaybot/modules/sticker_2.py
@@ -18,7 +18,10 @@ import os
 import random
 import textwrap
 import urllib
-
+from telethon import events
+from telethon.errors.rpcerrorlist import YouBlockedUserError
+from fridaybot import CMD_HELP, bot
+from fridaybot.utils import friday_on_cmd
 import emoji
 from fontTools.ttLib import TTFont
 from PIL import Image, ImageDraw, ImageFont, ImageOps
@@ -421,8 +424,13 @@ class Quote:
                 space += textfont.getsize(letter)[0]
 
 
-@borg.on(friday_on_cmd(pattern="q ?(.*)"))
+@borg.on(friday_on_cmd(pattern="q$"))
 async def create_sticker(message):
+    if event.fwd_from:
+        return
+    if not event.reply_to_msg_id:
+        await event.edit("```Reply to any user message.```")
+        return
     if not os.path.isdir(".tmp"):
         os.mkdir(".tmp", 0o755)
     await message.delete()
@@ -438,3 +446,44 @@ async def create_sticker(message):
     canvas.save(".tmp/sticker.webp")
     await message.respond(file=".tmp/sticker.webp")
     os.remove(".tmp/sticker.webp")
+
+@friday.on(friday_on_cmd(pattern=r"qbot(?: |$)(.*)"))
+async def _(event):
+    if event.fwd_from:
+        return
+    if not event.reply_to_msg_id:
+        await event.edit("```Reply to any user message.```")
+        return
+    reply_message = await event.get_reply_message()
+    chat = "@QuotLyBot"
+    reply_message.sender
+    if reply_message.sender.bot:
+        await event.edit("```Reply to actual users message.```")
+        return
+    await event.edit("```Making a Quote```")
+    async with bot.conversation(chat) as conv:
+        try:
+            response = conv.wait_event(
+                events.NewMessage(incoming=True, from_users=1031952739)
+            )
+            await bot.forward_messages(chat, reply_message)
+            response = await response
+        except YouBlockedUserError:
+            await event.reply("```Please unblock @QuotLyBot and try again```")
+            return
+        if response.text.startswith("Hi!"):
+            await event.edit(
+                "```Can you kindly disable your forward privacy settings for good?```"
+            )
+        else:
+            await event.delete()
+            await bot.forward_messages(event.chat_id, response.message)
+
+
+CMD_HELP.update(
+    {
+        "quotly": "**Quotly**\
+\n\n**Syntax : **`.qbot media <reply to some message>`\
+\n**Usage :** you will get quoted message."
+    }
+)
